rename map components
