+++ client/src/components/Maps/MobileMapPanel.js
@@ -1,7 +1,8 @@
 import React from "react"
 import PropTypes from "prop-types"
 import styled from "styled-components"
-import { media } from "../styles/theme/mixins"
+
+import { media } from "../../styles/theme/mixins"
 
 const MobileMapPanelStyles = styled.div`
   .mobile-trip-panel {
