+++ client/src/components/Maps/SingleTrip/Progress.js
@@ -2,6 +2,7 @@ import React from "react"
 import Styled from "styled-components"
 import { ProgressBar, Step } from "react-step-progress-bar"
 import PropTypes from "prop-types"
+
 import { media } from "../../../styles/theme/mixins"
 
 const ProgressPanel = Styled.div`
