+++ client/src/components/Maps/SingleTrip/TripInfo.js
