+++ client/src/components/Maps/SingleTrip/TripPanel.js
@@ -31,7 +31,7 @@ import * as util from "./mapUtil"
 import * as s from "./components"
 import { SERVER_URI } from "../../../config"
 import Waypoint from "./Waypoint"
-import MobileMapPanel from "../../MobileMapPanel"
+import MobileMapPanel from "../MobileMapPanel"
 import mapMarker from "../../icons/orange-marker.svg"
 import startMarker from "../../icons/green-marker.svg"
 import endMarker from "../../icons/black-marker.svg"
