+++ client/src/components/Maps/SingleTrip/components/index.js
