+++ client/src/components/Maps/SingleTrip/index.js
