+++ client/src/components/Maps/SingleTrip/mapUtil.js
