+++ client/src/components/NewTrip.js
@@ -1,5 +1,5 @@
 import React from "react"
-import CreateTrip from "../components/Maps/createTrip"
+import CreateTrip from "../components/Maps/CreateTrip"
 
 const NewTrip = () => {
   return <CreateTrip />
